move FileUtilities.read method to extension function within topl-core-base module
